separate EsriJSON from internal data structures
