+++ src/data/indexed/mod.rs
@@ -1,0 +1,8 @@
+mod indexed_data;
+pub use indexed_data::IndexedData;
+
+mod road_data_by_cwy;
+use road_data_by_cwy::RoadDataByCwy;
+
+#[cfg(test)]
+mod test_data;
