+++ src/data/mod.rs
@@ -1,6 +1,12 @@
-mod indexed_data;
-pub use indexed_data::IndexedData;
 
-pub mod index;
+/// This module contains a deserializer for a subset of the EsriJSON format.
+mod esri_json;
 
-pub mod esri_serde;
+/// This is the data we keep in memory and save/load from disk
+pub mod cached;
+
+/// This is a wrapper around the cached data which lets us query it really fast.
+mod indexed;
+pub use indexed::{
+    IndexedData
+};
