+++ src/filters/get_combined_filters.rs
@@ -4,7 +4,7 @@ use warp::{Filter, wrap_fn, filters::BoxedFilter, reply::Response, fs::File, Rep
 
 use crate::{data::IndexedData, settings::Settings};
 
-pub async fn load_data_and_combine_filters(settings:&Settings) -> Result<BoxedFilter<(Response,)>, Box<dyn Error>> {
+pub async fn load_data_and_get_combined_filters(settings:&Settings) -> Result<BoxedFilter<(Response,)>, Box<dyn Error>> {
 
     // Load data
     let indexed_data:Arc<_> = IndexedData::load(
@@ -36,6 +36,33 @@ pub async fn load_data_and_combine_filters(settings:&Settings) -> Result<BoxedFi
 }
 
 
+pub async fn get_combined_filters(settings:&Settings, indexed_data:Arc<IndexedData>) -> Result<BoxedFilter<(Response,)>, Box<dyn Error>> {
+
+    // define each "filter" (aka "route")  of the server
+    // each filter corresponds to a feature or capability
+    let filter_static_folder = warp::fs::dir(settings.NLR_STATIC_HTTP.clone());
+    let filter_show          = warp::path("show").and(filter_static_folder);
+    let filter_lines         = super::lines(indexed_data.clone());
+    let route_points         = super::points(indexed_data.clone());
+    let route_lines_batch    = super::lines_batch(indexed_data.clone());
+
+    // chain filters together into a single filter
+    let x = filter_show.map(|r:File| r.into_response()).or(
+        filter_lines
+        .or(route_points)
+        .or(
+            route_lines_batch
+            .with(warp::compression::gzip())
+        )
+        .recover(super::custom_rejection_handler)
+        .with(wrap_fn(super::echo_x_request_id))
+    ).unify();
+    Ok(x.boxed())
+}
+
+
+
+
 
 
 #[cfg(test)]
@@ -52,17 +79,26 @@ mod tests {
     /// we cannot avoid running this initialisation once for every
     /// individual test.
     macro_rules! setup_filter_for_testing {
-        () => {
+        () => {{
             // modified settings for testing
             //  - prevent saving the data file by providing an empty filepath
             //  - set data source url to download a small subset of data
             //    (The first 5-ish kilometres of H015)
-            load_data_and_combine_filters(&Settings {
+
+            let settings = Settings {
                 NLR_DATA_FILE: "".to_owned(), 
                 NLR_DATA_SOURCE_URL: "https://mrgis.mainroads.wa.gov.au/arcgis/rest/services/OpenData/RoadAssets_DataPortal/MapServer/17/query?where=ROAD%3D%27H015%27%20and%20END_SLK%3C5&outFields=ROAD,START_SLK,END_SLK,CWY&outSR=4326&f=json".to_owned(),
                 ..Settings::default()
-            }).await.unwrap()
-        };
+            };
+
+            let indexed_data:Arc<_> = IndexedData::load(
+                &settings.NLR_DATA_FILE,
+                &settings.NLR_DATA_SOURCE_URL,
+                &settings.NLR_FORCE_UPDATE_DATA
+            ).await.unwrap().into();
+
+            get_combined_filters(&settings, indexed_data).await.unwrap()
+        }};
     }
