+++ src/main.rs
@@ -1,24 +1,23 @@
-use std::{net::SocketAddr, error::Error};
-use clap::Parser;
-
 mod helpers;
 mod filters;
 mod data;
 mod settings;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn Error>> {
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
     
-    // Read settings from command line (or environment variables)
-    // see run with --help to see nice instructions generated by clap
-    let settings = settings::Settings::parse();
+    let settings = settings::Settings::get_from_cli_or_env();
 
-    let filter = filters::load_data_and_combine_filters(&settings).await?;
+    let indexed_data = data::IndexedData::load(
+        &settings.NLR_DATA_FILE,
+        &settings.NLR_DATA_SOURCE_URL,
+        &settings.NLR_FORCE_UPDATE_DATA
+    ).await?;
 
-    // Serve
-    let address:SocketAddr = (settings.NLR_ADDR, settings.NLR_PORT).into();
-    println!("Serving at {:?}", address);
-    warp::serve(filter).run(address).await;
+    let filter = filters::get_combined_filters(&settings, indexed_data.into()).await?;
+
+    println!("Serving at {:?}", settings.get_socket_address());
+    warp::serve(filter).run(settings.get_socket_address()).await;
 
     Ok(())
 }
